+++ R/tidytable_package.R
@@ -7,6 +7,7 @@
 #' @importFrom rlang `%||%`
 #' @importFrom stats as.formula na.omit setNames
 #' @importFrom utils capture.output head tail getFromNamespace
+#' @importFrom vctrs vec_assert vec_cast vec_ptype_common
 NULL
 
 globalVariables(c("data", ".","..select_vars", ".count", "na_index", ".new_col",